R6 class for a generic normal model

R6 class for a generic normal model

R6 class for a generic normal model

Super classes

normalblockr::normal_models -> normalblockr::NB -> NB_zi_fixed_Q

Public fields

fixed_tau

whether tau should be fixed at clustering_init during optimization, useful for stability selection

zeros

indicator matrix of zeros in Y

Active bindings

nb_param

number of parameters in the model

var_par

a list with variational parameters

model_par

a list with model parameters: B (covariates), dm1 (species variance), OmegaQ (blocks precision matrix), kappa (zero-inflation probabilities)

entropy

Entropy of the variational distribution when applicable

fitted

Y values predicted by the model Y values predicted by the model

who_am_I

a method to print what model is being fitted

Methods

Public methods

NB_zi_fixed_Q$new()

NB_zi_fixed_Q$clone()

Inherited methods

normalblockr::normal_models$optimize()

normalblockr::normal_models$plot_loglik()

normalblockr::normal_models$update()

normalblockr::NB$latent_network()

normalblockr::NB$merge()

normalblockr::NB$plot_network()

normalblockr::NB$split()

Method new()

Create a new [`NB_zi_fixed_Q`] object.

Usage

NB_zi_fixed_Q$new(data, Q, sparsity = 0, control = NB_control())

Arguments

data

object of normal_data class, with responses and design matrix

Q

required number of groups

sparsity

to apply on variance matrix when calling GLASSO

control

structured list of more specific parameters

Returns

A new [`NB_zi_fixed_Q`] object

Method clone()

The objects of this class are cloneable with this method.

Usage

NB_zi_fixed_Q$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.