R6 class for a normal-block model with fixed number of clusters (but unknown clustering).

R6 class for a normal-block model with fixed number of clusters (but unknown clustering).

Super class

NormalBlockBase -> NormalBlockUnknownClusters

Public fields

fixed_tau

whether tau should be fixed at clustering_init during optimization, useful for stability selection

Active bindings

model_par

a list with the matrices of the model parameters: B (covariates), dm1 (species variance), Omegaq (groups precision matrix))

nb_param

number of parameters in the model

var_par

a list with the matrices of the variational parameters: M (means), S (variances), tau (posterior group probabilities)

entropy

Entropy of the conditional distribution

fitted

Y values predicted by the model, in Y's original units

who_am_I

a method to print what model is being fitted

Methods

Public methods

NormalBlockUnknownClusters$new()

NormalBlockUnknownClusters$clone()

Inherited methods

NormalBlockBase$best_of_inits()

NormalBlockBase$candidates_merge()

NormalBlockBase$candidates_split()

NormalBlockBase$latent_network()

NormalBlockBase$merge()

NormalBlockBase$optimize()

NormalBlockBase$plot()

NormalBlockBase$plot_loglik()

NormalBlockBase$plot_network()

NormalBlockBase$predict()

NormalBlockBase$print()

NormalBlockBase$split()

NormalBlockBase$update()

NormalBlockBase$warm_start_from()

NormalBlockUnknownClusters$new()

Create a new [`NormalBlockUnknownClusters`] object.

Usage

NormalBlockUnknownClusters$new(data, q, sparsity = 0, control = NB_control())

Arguments

data

contains the matrix of responses (Y) and the design matrix (X).

q

required number of groups

sparsity

sparsity penalty to add on blocks precision matrix for sparsity

control

structured list for specific parameters

Returns

A new [`NormalBlockUnknownClusters`] object

NormalBlockUnknownClusters$clone()

The objects of this class are cloneable with this method.

Usage

NormalBlockUnknownClusters$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.