R6 class for a Zero-Inflated normal-block model with a known clustering.

R6 class for a Zero-Inflated normal-block model with a known clustering.

R6 class for a Zero-Inflated normal-block model with a known clustering.

Super class

normalblockr::NormalBlockVarBase -> ZINormalBlockVarKnownClusters

Active bindings

posterior_par

a list with the parameters of posterior distribution W | Y

entropy

Entropy of the conditional distribution

nb_param

number of parameters in the model

model_par

a list with model parameters: B (covariates), dm1 (species variance), Omegaq (groups precision matrix), kappa (zero-inflation probabilities)

fitted

Y values predicted by the model, in Y's original units

who_am_I

a method to print what model is being fitted

Methods

Public methods

ZINormalBlockVarKnownClusters$new()

ZINormalBlockVarKnownClusters$clone()

Inherited methods

normalblockr::NormalBlockVarBase$best_of_inits()

normalblockr::NormalBlockVarBase$candidates_merge()

normalblockr::NormalBlockVarBase$candidates_split()

normalblockr::NormalBlockVarBase$latent_network()

normalblockr::NormalBlockVarBase$merge()

normalblockr::NormalBlockVarBase$optimize()

normalblockr::NormalBlockVarBase$plot()

normalblockr::NormalBlockVarBase$plot_loglik()

normalblockr::NormalBlockVarBase$plot_network()

normalblockr::NormalBlockVarBase$predict()

normalblockr::NormalBlockVarBase$print()

normalblockr::NormalBlockVarBase$split()

normalblockr::NormalBlockVarBase$update()

normalblockr::NormalBlockVarBase$warm_start_from()

Method new()

Create a new [`ZINormalBlockVarKnownClusters`] object.

Usage

ZINormalBlockVarKnownClusters$new(
  data,
  C,
  sparsity = 0,
  control = NB_control()
)

Arguments

data

object of NormalBlockData class, with responses and design matrix

C

clustering matrix C_jk = 1 if species j belongs to cluster k

sparsity

to apply on variance matrix when calling GLASSO

control

structured list of more specific parameters, to generate with NB_control

Returns

A new [`ZINormalBlockVarKnownClusters`] object

Method clone()

The objects of this class are cloneable with this method.

Usage

ZINormalBlockVarKnownClusters$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.