R6 class for a generic normal_diag_zi model

R6 class for a generic normal_diag_zi model

R6 class for a generic normal_diag_zi model

Super class

normalblockr::normal_models -> normal_diag_zi

Public fields

zeros

indicator matrix of zeros in Y

Active bindings

nb_param

number of parameters in the model

model_par

a list with the matrices of the model parameters: B (regression coefficients), dm1 (species variance), kappa (zero-inflation probas)), rho (zero-inflation posterior proba)

fitted

Y values predicted by the model Y values predicted by the model

who_am_I

a method to print what model is being fitted

Methods

Public methods

normal_diag_zi$new()

normal_diag_zi$clone()

Inherited methods

normalblockr::normal_models$optimize()

normalblockr::normal_models$plot_loglik()

normalblockr::normal_models$update()

Method new()

Create a new [`normal_diag_zi`] object.

Usage

normal_diag_zi$new(data)

Arguments

data

contains the matrix of responses (Y) and the design matrix (X).

Returns

A new [`normal_diag_zi`] object

Method clone()

The objects of this class are cloneable with this method.

Usage

normal_diag_zi$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.