R6 class for a generic normal_fixed_sparsity model

R6 class for a generic normal_fixed_sparsity model

R6 class for a generic normal_fixed_sparsity model

Public fields

data

object of normal_data class, with responses and design matrix

Active bindings

n

number of samples

p

number of responses per sample

d

number of variables (dimensions in X)

model_par

a list with the matrices of the model parameters: B (covariates), dm1 (species variance), OmegaQ (groups precision matrix))

loglik

(or its variational lower bound)

nb_param

number of parameters in the model

entropy

Entropy of the variational distribution when applicable

deviance

(or its variational lower bound)

BIC

(or its variational lower bound)

ICL

variational lower bound of the ICL

criteria

a vector with loglik, BIC and number of parameters

objective

evolution of the objective function during (V)EM algorithm

Methods

Public methods

normal_models$new()

normal_models$update()

normal_models$optimize()

normal_models$plot_loglik()

normal_models$clone()

Method new()

Create a new [`normal_models`] object.

Usage

normal_models$new(data, control = NB_control())

Arguments

data

object of normal_data class, with responses and design matrix

control

structured list of more specific parameters, to generate with NB_control (useful only for NB objects)

Returns

A new [`nb_fixed`] object

Method update()

Update a [`normal_models`] object

All possible parameters of the child classes

Usage

normal_models$update(
  B = NA,
  dm1 = NA,
  C = NA,
  OmegaQ = NA,
  gamma = NA,
  mu = NA,
  kappa = NA,
  rho = NA,
  alpha = NA,
  M = NA,
  S = NA,
  ll_list = NA
)

Arguments

B

regression matrix

dm1

diagonal vector of inverse variance matrix (variables level)

C

the matrix of groups memberships (posterior probabilities)

OmegaQ

groups inverse variance matrix

gamma

variance of posterior distribution of W

mu

mean for posterior distribution of W

kappa

vector of zero-inflation probabilities

rho

posterior probabilities of zero-inflation

alpha

vector of groups probabilities

M

variational mean for posterior distribution of W

S

variational diagonal of variances for posterior distribution of W

ll_list

list of log-lik (elbo) values

Returns

Update the current [`normal`] object

Method optimize()

calls optimization (EM or heuristic) and updates relevant fields

Usage

normal_models$optimize(control = list(niter = 100, threshold = 1e-04))

Arguments

control

a list for controlling the optimization proces

Returns

optimizes the model and updates its parameters

Method plot_loglik()

plots log-likelihood values during model optimization

Usage

normal_models$plot_loglik(type = "b", log = "", neg = FALSE)

Arguments

type

char for line type (see plot.default)

log

char for logarithmic axes (see plot.default)

neg

boolean plot negative log-likelihood (useful when log="y")

Method clone()

The objects of this class are cloneable with this method.

Usage

normal_models$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.